Moves test failure event cross version tests to r900
Moves test failure progress event cross version tests related to Gradle 9.0 features from r83 to a dedicated r900 test suite.
